prompt for delete profile

+++ gui_dlg_profiles.py
@@ -86,8 +86,10 @@ class ProfilesDialog(gui_base_dialog.PlanerBaseModalDialog):
             messagebox.showinfo(self.T("caption_title"),
                                 self.T("delete_error_curr_profile"))  # Aktywny profil nie może zostać usunięty.
         else:
-            # TODO: Okienko powierdzenia usunięcia profilu.
-            self.state.delete_profile(profile_id)
+            ynd = gui_base_dialog.YesNoDialog(self, self.T("delete_profile_prompt"), title=None)
+            self.top.wait_window(ynd.top)
+            if ynd.yes:
+                self.state.delete_profile(profile_id)
         self.refresh()
 
     def on_new_profile(self):
